Fix babel parsing plugins (typescript)
I was mystified as to why stylelint stopped working when I switched from .jsx to .tsx, and it turns out it was this. It was overriding that set of plugins with Babel's automatically detected ones, and so would fail if you were using tsc, or in my case, had babel typescript installed in a gatsby plugin, which babel missed.

I imagine there are some other problems this might solve too though. Please give feedback or edit yourself, I'm not able to thoroughly test this.